organizing routes
created an index.html file and put it an a public folder along with a folder to keep all of our routes in. i successfully created an html route and exported it to the server file. when we run the server now on localhost it gets the html file that juan created from its own html route file and serves the static index file from the router. i put the auth route file that marcio created in the routes folder as well so now we can begin creating more route files and to keep everything organized